Resolved data calculation issues and verified expected output

+++ src/main/java/models/Opportunity.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public class Opportunity {
     public double sharpeRatio;
-    public Map<String, Float> weights;
+    public Map<String, Double> weights;
     public double portfolioReturn;
     public double portfolioVariance;
     public double portfolioStDeviation;
